requirements.txt
Flask
flask-cors
python-dotenv
roboflow
numpy
opencv-python

